fix: Improve tiapp sdk-version handling
